Added iron and rock blocks, arrows now are better predicted

+++ MiningGameserver/Packets/Packet2SCCreateProjectile.cs
@@ -2,14 +2,14 @@
 {
     public class Packet2SCCreateProjectile : Packet
     {
-        public Packet2SCCreateProjectile(byte ProjectileID, byte projectileType, short X, short Y, float angle, byte strength, byte owner)
+        public Packet2SCCreateProjectile(byte ProjectileID, byte projectileType, short X, short Y, short angle, byte strength, byte owner)
         {
             writeByte(2);
             writeByte(projectileType);
             writeByte(ProjectileID);
             writeShort(X);
             writeShort(Y);
-            writeFloat(angle);
+            writeShort(angle);
             writeByte(strength);
             writeByte(owner);
         }
